Implemented userCheckIn in OccupancyManager, and implemented tests

+++ src/main/java/com/studyspaces/spacefinder/repository/RealTimeOccupancyRepository.java
@@ -27,7 +27,7 @@ public interface RealTimeOccupancyRepository
      * Retrieve all documents that contain records after a given timestamp.
      */
     @Query("{ 'records.timestamp': { $gt: ?0 } }")
-    List<RoomOccupancyRecord> findWithRecordsAfterTimestamp(int timestamp);
+    List<RoomOccupancyRecord> findWithRecordsAfterTimestamp(long timestamp);
 
     /**
      * Retrieve a specific document and filter its records by timestamp.
